Modified interface value type int -> long to satisfy number size in Day2 Challenge

+++ src/main/java/software/aoc/challenges/Solver.java
@@ -1,5 +1,5 @@
 package software.aoc.challenges;
 
 public interface Solver {
-    int solve(String input);
+    long solve(String input);
 }
